+++ R/rectangle.R
@@ -61,15 +61,18 @@
 #'   particular column. If a named list is provided, the default for any
 #'   unspecified columns is `TRUE`.
 #' @param .ptype,ptype Optionally, a named list of prototypes declaring the
-#'   desired output type of each component. Use this argument if you want to
-#'   check that each element has the type you expect when simplifying.
+#'   desired output type of each component. Alternatively, a single empty
+#'   prototype can be supplied, which will be applied to all components. Use
+#'   this argument if you want to check that each element has the type you
+#'   expect when simplifying.
 #'
 #'   If a `ptype` has been specified, but `simplify = FALSE` or simplification
 #'   isn't possible, then a [list-of][vctrs::list_of()] column will be returned
 #'   and each element will have type `ptype`.
 #' @param .transform,transform Optionally, a named list of transformation
-#'   functions applied to each component. Use this function if you want
-#'   to transform or parse individual elements as they are extracted.
+#'   functions applied to each component. Alternatively, a single function can
+#'   be supplied, which will be applied to all components. Use this argument if
+#'   you want to transform or parse individual elements as they are extracted.
 #'
 #'   When both `ptype` and `transform` are supplied, the `transform` is applied
 #'   before the `ptype`.
@@ -175,8 +178,8 @@ hoist <- function(.data,
                   ...,
                   .remove = TRUE,
                   .simplify = TRUE,
-                  .ptype = list(),
-                  .transform = list()) {
+                  .ptype = NULL,
+                  .transform = NULL) {
   if (!is.data.frame(.data)) {
     abort("`.data` must be a data frame.")
   }
@@ -324,8 +327,8 @@ unnest_longer <- function(data,
                           indices_include = NULL,
                           names_repair = "check_unique",
                           simplify = TRUE,
-                          ptype = list(),
-                          transform = list()) {
+                          ptype = NULL,
+                          transform = NULL) {
   if (!is.data.frame(data)) {
     abort("`data` must be a data frame.")
   }
@@ -558,8 +561,8 @@ unnest_wider <- function(data,
                          simplify = TRUE,
                          strict = FALSE,
                          names_repair = "check_unique",
-                         ptype = list(),
-                         transform = list()) {
+                         ptype = NULL,
+                         transform = NULL) {
   if (!is.data.frame(data)) {
     abort("`data` must be a data frame.")
   }
@@ -766,30 +769,13 @@ guess_dir <- function(x, col) {
 
 df_simplify <- function(x,
                         ...,
-                        ptype = list(),
-                        transform = list(),
+                        ptype = NULL,
+                        transform = NULL,
                         simplify = TRUE) {
   ellipsis::check_dots_empty()
 
-  if (!vec_is_list(ptype)) {
-    abort("`ptype` must be a list.")
-  }
-  if (length(ptype) > 0L && !is_named(ptype)) {
-    abort("All elements of `ptype` must be named.")
-  }
-  if (vec_duplicate_any(names(ptype))) {
-    abort("The names of `ptype` must be unique.")
-  }
-
-  if (!vec_is_list(transform)) {
-    abort("`transform` must be a list.")
-  }
-  if (length(transform) > 0L && !is_named(transform)) {
-    abort("All elements of `transform` must be named.")
-  }
-  if (vec_duplicate_any(names(transform))) {
-    abort("The names of `transform` must be unique.")
-  }
+  ptype <- check_list_of_ptypes(ptype, names(x), "ptype")
+  transform <- check_list_of_functions(transform, names(x), "transform")
 
   if (is_bool(simplify)) {
     simplify_default <- simplify